+++ doc/package-search-index.js
@@ -1,1 +1,1 @@
-packageSearchIndex = [{"l":"All Packages","u":"allpackages-index.html"},{"l":"java_cp1_javier_guerra"}];updateSearchResults();
+packageSearchIndex = [{"l":"All Packages","u":"allpackages-index.html"},{"l":"java_cp1_javier_guerra"},{"l":"java_cp1_javier_guerra.entities"},{"l":"java_cp1_javier_guerra.services"},{"l":"java_cp1_javier_guerra.utils"}];updateSearchResults();