+++ src/java_cp1_javier_guerra/entities/Media.java
@@ -1,7 +1,5 @@
 package java_cp1_javier_guerra.entities;
 
-import java_cp1_javier_guerra.interfaces.MediaInterface;
-
 import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;