+++ src/java_cp1_javier_guerra/entities/MediaInterface.java
@@ -1,4 +1,4 @@
-package java_cp1_javier_guerra.interfaces;
+package java_cp1_javier_guerra.entities;
 
 import java_cp1_javier_guerra.entities.Song;
 import java_cp1_javier_guerra.entities.MediaFormat;